No explicit _index in bulk requests
Make the bulk indexing compatible with ES clusters having the
rest.action.multi.allow_explicit_index set to false.

This is of particular interest in AWS ES Service clusters since
that setting is the only way to configure per-index access control.

The commit introduces a division of the batch into one bulk request
per index, and POSTs the data to /indexname/_bulk endpoint instead
of the /_bulk endpoint.