6.4.0 version release  (#96)
Upgrade to 6.4.0 version:

What's new

*Template management for Elasticsearch*
Now, the plugin support not only the newest version template but also can adjust template according users' es version to set corresponding template。

*Retry Policy*

The retry policy has changed after 6.4.0 release. This plugin uses the Elasticsearch bulk API to optimize its imports into Elasticsearch. These requests may experience either partial or total failures. The bulk API sends batches of requests to an HTTPS endpoint. Error codes for the HTTPS request are handled differently than error codes for individual documents.
HTTPS requests to the bulk API are expected to return a 200 response code. All other response codes are retried indefinitely.
The following document errors are handled as follows:

* 400 and 404 errors are sent to the dead letter queue (DLQ), if enabled. If a DLQ is not enabled, a log message will be emitted, and the event will be dropped. See DLQ Policy for more info.
* 409 errors (conflict) are logged as a warning and dropped.



*DLQ Policy*

Mapping (404) errors from Elasticsearch can lead to data loss. Unfortunately mapping errors cannot be handled without human intervention and without looking at the field that caused the mapping mismatch. If the DLQ is enabled, the original events causing the mapping errors are stored in a file that can be processed at a later time. Often times, the offending field can be removed and re-indexed to Elasticsearch. If the DLQ is not enabled, and a mapping error happens, the problem is logged as a warning, and the event is dropped. .


*Bath Size*

After 6.4.0 ,we will not support *flush_size*  and *idle_flush_time* any more, this plugin attempts to send batches of events as a single request. However, if a request exceeds 20MB we will break it up until multiple batch requests. If a single document exceeds 20MB it will be sent as a single request.